+++ sei-ibc-go/CHANGELOG.md
@@ -39,6 +39,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 ### Dependencies
 
 * [\#1525](https://github.com/cosmos/ibc-go/pull/1525) Bump SDK version to v0.45.5
+* [\#1627](https://github.com/cosmos/ibc-go/pull/1627) Bump Go version to 1.18
 
 ### API Breaking